truffle-occurrence-downloader/requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.64.0

# Data processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Optional geospatial dependencies (install with pip install -e .[geospatial])
# geopandas>=0.13.0
# shapely>=2.0.0
# fiona>=1.8.0

# Development dependencies (install with pip install -e .[dev])
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0
# mypy>=1.0.0